Detail screen

+++ app/src/main/java/com/bishal/lazyreader/model/ListPriceX.kt
@@ -1,6 +1,6 @@
 package com.bishal.lazyreader.model
 
 data class ListPriceX(
-    val amountInMicros: Int,
+    val amountInMicros: Long,
     val currencyCode: String
 )
